requirements.txt
backports.functools-lru-cache==1.6.4
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.7
capstone==5.0.0rc2
certifi==2021.10.8
cffi==1.15.1
chardet==4.0.0
colored-traceback==0.3.0
cryptography==3.3.2
decorator==4.4.2
enum34==1.1.10
futures==3.3.0
idna==2.10
intervaltree==3.1.0
ipaddress==1.0.23
ipdb==0.13.9
ipython==5.10.0
ipython-genutils==0.2.0
lxml==4.9.1
Mako==1.1.6
MarkupSafe==1.1.1
opcua==0.98.13
packaging==20.9
paramiko==2.11.0
pathlib2==2.3.7.post1
pexpect==4.8.0
pickleshare==0.7.5
plumbum==1.7.2
prompt-toolkit==1.0.18
psutil==5.9.1
ptyprocess==0.7.0
pwntools==4.8.0
pycparser==2.21
pyelftools==0.28
Pygments==2.5.2
pymodbus==2.5.3
pytz==2022.1
requests==2.27.1
ROPGadget==6.8
rpyc==4.1.5
scandir==1.10.0
simplegeneric==0.8.1
six==1.16.0
sortedcontainers==2.4.0
toml==0.10.2
traitlets==4.3.3
trollius==2.2.1
typing==3.10.0.0
unicorn==2.0.0
urllib3==1.26.10
wcwidth==0.2.5
